Docs: Change tabs shortcode example to permit unsafe=false (#689)
The syntax of `{{% code %}}` only expects Markdown inside. That's why it
triggers unsafe warning, if used as a shortcode template.

Instead the only the top table shortcode should be used as
`{{< tabs >}}` to include the HTML, while the user may use only
`{{% tab "Name" %}}` inside and be happily restricted to Markdown.

https://gohugo.io/methods/page/rendershortcodes/#shortcode-notation
https://github.com/InfosecForActivistsTeam/infosec-activists/pull/45